Add of Delete Button Functionality to feature/addToCartLists

+++ ESEWA MARKET/Home/Cart/View/CartViewController.swift
@@ -16,14 +16,14 @@ class CartViewController: UIViewController, AddItemToCartProtocol {
     
     // MARK: - Add Custom FooterView
     
-     var cartFooterView: UIView = {
+    let cartFooterView: UIView = {
         let footerView = UIView()
         footerView.translatesAutoresizingMaskIntoConstraints = false
         footerView.backgroundColor = .white
         return footerView
     }()
     
-    var itemCheckOutTotal: UILabel = {
+    let itemCheckOutTotal: UILabel = {
         let total = UILabel()
         total.text = "Checkout Total"
         total.translatesAutoresizingMaskIntoConstraints = false
@@ -33,9 +33,9 @@ class CartViewController: UIViewController, AddItemToCartProtocol {
         return total
     }()
     
-    var itemTotalPriceLabel: UILabel = {
+    let itemTotalPriceLabel: UILabel = {
         let totalPrice = UILabel()
-        totalPrice.text = "Rs.52,500"
+//        totalPrice.text = "Rs.52,500"
         totalPrice.translatesAutoresizingMaskIntoConstraints = false
         totalPrice.textAlignment = .left
         totalPrice.font = .systemFont(ofSize: 16, weight: .medium)
@@ -43,7 +43,7 @@ class CartViewController: UIViewController, AddItemToCartProtocol {
         return totalPrice
     }()
     
-    var itemCheckOutButton: UIButton = {
+    let itemCheckOutButton: UIButton = {
         let checkOutButton = UIButton()
         checkOutButton.translatesAutoresizingMaskIntoConstraints = false
         checkOutButton.setTitle("CHECKOUT", for: .normal)
@@ -55,7 +55,7 @@ class CartViewController: UIViewController, AddItemToCartProtocol {
     
     // MARK: - Add a table View
     
-    var tableView: UITableView = {
+    let tableView: UITableView = {
         let table = UITableView()
         //table.backgroundColor = .clear//.systemFill
         table.register(CustomTableViewCell.self, forCellReuseIdentifier: CustomTableViewCell.identifier)
@@ -76,7 +76,7 @@ class CartViewController: UIViewController, AddItemToCartProtocol {
         let backButton = UIBarButtonItem(image: UIImage(systemName: "chevron.backward"), style: .plain, target: self, action: #selector(backButtonPressed))
         button.tintColor = .black
         backButton.tintColor = .black
-        self.navigationController?.popViewController(animated: true)
+        //        self.navigationController?.popViewController(animated: true)
         navigationItem.rightBarButtonItem = button
         navigationItem.leftBarButtonItem = backButton
         navigationItem.title = "My Cart"
@@ -97,7 +97,7 @@ class CartViewController: UIViewController, AddItemToCartProtocol {
         updateTotalPriceLabel()
         
     }
-   
+    
     private func setupTableView() {
         tableView.delegate = self
         tableView.dataSource = self
@@ -112,19 +112,19 @@ class CartViewController: UIViewController, AddItemToCartProtocol {
         self.model = model
         tableView.reloadData()
     }
-
+    
     @objc func backButtonPressed() {
         self.navigationController?.popViewController(animated: true)
-     
+        
     }
     
     func updateTotalPriceLabel() {
         itemTotalPriceLabel.text = "Rs.\(totalPrice)"
-
+        
     }
     private func setupFooterView() {
-    
-            // MARK: - Set Constraint of footerView
+        
+        // MARK: - Set Constraint of footerView
         
         NSLayoutConstraint.activate([
             cartFooterView.leadingAnchor.constraint(equalTo: self.tableView.leadingAnchor, constant: 0),
@@ -154,7 +154,6 @@ class CartViewController: UIViewController, AddItemToCartProtocol {
     }
     
 }
-
 extension CartViewController: UITableViewDelegate {
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         print("clicked cell at \(indexPath.row)")
@@ -165,47 +164,57 @@ extension CartViewController: UITableViewDelegate {
     }
     
     func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCell.EditingStyle, forRowAt indexPath: IndexPath) {
+        
+        if editingStyle == .delete {
+            let vc = DeletePopUpViewController()
+            vc.onDeleteClicked = { deletedMessage in
+                self.presenter?.cartItemsList.remove(at: indexPath.row)
+                tableView.deleteRows(at: [indexPath], with: .middle)
+                vc.dismiss(animated: true) {
+                    self.tableView.reloadData()
+                    print(deletedMessage)
+                }
+            }
 
-        let vc = DeletePopUpViewController()
-        vc.modalPresentationStyle = .popover
-        vc.preferredContentSize.height = 50
-        self.present(vc, animated: true)
+            vc.modalPresentationStyle = .popover
+            vc.preferredContentSize.height = 50
+            self.present(vc, animated: true, completion: nil)
+        }
+    }
+    
+}
 
-
+extension CartViewController: UITableViewDataSource {
+    func numberOfSections(in tableView: UITableView) -> Int {
+        return 1
+    }
+    
+    func tableView(_ tableView: UITableView, titleForHeaderInSection section: Int) -> String? {
+        return "Items(\(presenter?.cartItemsList.count))"
+        
+    }
+    
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return presenter?.cartItemsList.count ?? 1
+    }
+    
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        let cell = tableView.dequeueReusableCell(withIdentifier: CustomTableViewCell.identifier, for: indexPath) as! CustomTableViewCell
+        
+        let item = model[indexPath.row]
+        cell.configure(with: item)
+        cell.backgroundColor = .clear
+        
+        cell.countChanged = { price in
+            self.totalPrice = (price + price)
+            self.updateTotalPriceLabel()
+        }
+        return cell
+    }
+    
+    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
+        return 150
     }
 }
-          
-    extension CartViewController: UITableViewDataSource {
-        func numberOfSections(in tableView: UITableView) -> Int {
-            return 1
-        }
-        
-        func tableView(_ tableView: UITableView, titleForHeaderInSection section: Int) -> String? {
-            return "Items(\(model.count))"
-            
-        }
-        
-        func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-            return model.count
-        }
-        
-        func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-            let cell = tableView.dequeueReusableCell(withIdentifier: CustomTableViewCell.identifier, for: indexPath) as! CustomTableViewCell
-            
-            let item = model[indexPath.row]
-            cell.configure(with: item)
-            cell.backgroundColor = .clear
-            cell.countChanged = { price in
-                self.totalPrice += price
-                self.updateTotalPriceLabel()
-            }
-            return cell
-        }
-        
-        func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-            return 150
-        }
-}
-    
-    
+
